LayoutTests/inspector/editor/highlighter-chunk-limit-expected.txt

This test checks highlighting with specific highlightChunkLimit.

Highlighter results
1 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
2 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
3 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
4 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
5 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
6 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
7 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
8 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
9 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
10 : this.field = this.field = "foo"; /* comment here */ this.otherField = 12; // end of line : javascript-keyword[0-4] javascript-ident[5-10] javascript-keyword[13-17] javascript-ident[18-23] javascript-string[26-31] javascript-comment[33-51] javascript-keyword[52-56] javascript-ident[57-67] javascript-number[70-72] javascript-comment[74-88]
11 :  :
